Fixed time model for improved accuracy on the prediction. Added mechanism to improve the reception of multiple packets in one wake-up.

+++ core/net/mac/EM-MAC/Neighbors_list.h
@@ -11,7 +11,7 @@ typedef struct {
 	unsigned int 	last_seed;
 	uint8_t			last_channel;
 	unsigned int 	wake_time_seconds;
-	long int 		n;                       // time difference in tics
+	int 			n;                       // time difference in tics
 	unsigned int 	blacklist;
 	unsigned short	consecutive_failed_tx;
 } neighbor_state;
